Add string formatting for addresses and rework printing
Prior to this commit, there was one function for printing virtual
addresses to the standard output and one for physical addresses.  While
this was the simplest possible solution, it could not be used to print
addresses together with surrounding text in a single thread-safe
`printf()` call.

This commit keeps those simple functions for printing addresses (*) and
adds two new features to the library that allow address strings to be
embedded in a `printf()` call:  The `sprint_{phys,virt}_addr()` function
writes an address into a character string buffer.  That buffer needs to
be allocated before calling the function according to the value of
`{PHYS,VIRT}_ADDR_STRLEN`.

(*) The `print_{phys,virt}_addr` functions now use heap memory instead
of stack memory, meaning they can return `-ENOMEM` if the allocation on
heap failed.